fixed some errors & jwt auth (in progress)
