+++ client/src/api/axios.ts
@@ -1,0 +1,6 @@
+import axios from "axios";
+
+export default axios.create({
+  baseURL: import.meta.env.VITE_REACT_BASE_URL,
+  withCredentials: true,
+});
