--- server/src/middleware/auth.ts
@@ -1,27 +1,0 @@
-import jwt from "jsonwebtoken";
-import { Request, Response, NextFunction } from "express";
-
-export const verifyToken = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
-    let token = req.header("Authorization");
-
-    if (!token) {
-      return res.status(403).send("Access denied");
-    }
-    if (token.startsWith("Bearer ")) {
-      token = token.slice(7, token.length).trimStart();
-    }
-
-    const verified = jwt.verify(token, process.env.JWT_SECRET);
-    req.payload = verified;
-    next();
-  } catch (error) {
-    if (error instanceof Error) {
-      res.status(500).json({ error: error.message });
-    }
-  }
-};
