+++ server/src/server.ts
@@ -12,8 +12,6 @@ export default class Server {
     const corsOptions: CorsOptions = {
       origin: "http://localhost:5173",
       credentials: true,
-      methods: "GET, POST, PUT, DELETE",
-      allowedHeaders: "Content-Type, *",
     };
 
     dotenv.config();
